requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
